+++ twimi-app/src/lib.rs
@@ -3,11 +3,11 @@ pub mod server;
 
 use shaku::module;
 
+use infrastructure::services::{Argon2PasswordHasher, Argon2PasswordVerifier};
 use infrastructure::{
     repositories::{health_check::HealthCheckRepositoryImpl, users::UsersRepositoryImpl},
     MySqlDatabase,
 };
-use server::helpers::{Argon2PasswordHasher, Argon2PasswordVerifier};
 use twimi_core::usecases::{
     health_check::HealthCheckUseCase, login::LoginUseCase, signup::SignUpUseCase,
 };
