+++ twimi-app/src/server/mod.rs
@@ -1,4 +1,3 @@
-pub mod helpers;
 pub mod models;
 pub mod routes;
 pub mod startup;
